extract Event::Cloner.
